# openapi.json
{"openapi": "3.1.0", "info": {"title": "Your API's Title", "version": "1.0.0"}, "paths": {"/login": {"post": {"summary": "Login", "operationId": "login_login_post", "parameters": [{"name": "password", "in": "query", "required": true, "schema": {"type": "string", "title": "Password"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/": {"get": {"summary": "Welcome", "operationId": "welcome__get", "responses": {"200": {"description": "Successful Response", "content": {"text/html": {"schema": {"type": "string"}}}}}}}, "/producao": {"get": {"summary": "Get Producao", "operationId": "get_producao_producao_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProducaoResponse"}}}}}, "security": [{"HTTPBearer": []}]}}, "/processamento": {"get": {"summary": "Get Processamento", "operationId": "get_processamento_processamento_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProcessamentoResponse"}}}}}, "security": [{"HTTPBearer": []}]}}, "/comercializacao": {"get": {"summary": "Get Comercializacao", "operationId": "get_comercializacao_comercializacao_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ComercializacaoResponse"}}}}}, "security": [{"HTTPBearer": []}]}}, "/importacao": {"get": {"summary": "Get Importacao", "operationId": "get_importacao_importacao_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ImportacaoResponse"}}}}}, "security": [{"HTTPBearer": []}]}}, "/exportacao": {"get": {"summary": "Get Exportacao", "operationId": "get_exportacao_exportacao_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ExportacaoResponse"}}}}}, "security": [{"HTTPBearer": []}]}}}, "components": {"schemas": {"ComercializacaoResponse": {"properties": {"comercializacao": {"items": {}, "type": "array", "title": "Comercializacao"}}, "type": "object", "required": ["comercializacao"], "title": "ComercializacaoResponse"}, "ExportacaoResponse": {"properties": {"exportacao": {"items": {}, "type": "array", "title": "Exportacao"}}, "type": "object", "required": ["exportacao"], "title": "ExportacaoResponse"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "ImportacaoResponse": {"properties": {"importacao": {"items": {}, "type": "array", "title": "Importacao"}}, "type": "object", "required": ["importacao"], "title": "ImportacaoResponse"}, "ProcessamentoResponse": {"properties": {"processamento": {"items": {}, "type": "array", "title": "Processamento"}}, "type": "object", "required": ["processamento"], "title": "ProcessamentoResponse"}, "ProducaoResponse": {"properties": {"producao": {"items": {}, "type": "array", "title": "Producao"}}, "type": "object", "required": ["producao"], "title": "ProducaoResponse"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}, "securitySchemes": {"HTTPBearer": {"type": "http", "scheme": "bearer"}}}}